backend/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
Flask_JWT_Extended==4.6.0
flask_mail==0.10.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
itsdangerous==2.2.0
Pillow==10.4.0
qrcode==7.4.2
Werkzeug==3.0.3
WTForms==3.1.2
